Keep profile module within size budget
